pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "LeLab"
version = "0.1.0"
description = "LeRobot Lab - A web interface for robotics with lerobot integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "websockets>=15.0.1",
    "uvicorn>=0.24.0",
    "python-socketio>=5.11.0",
    "lerobot @ git+https://github.com/huggingface/lerobot.git",
]

[project.scripts]
lelab = "scripts.backend:main"
"lelab-fullstack" = "scripts.fullstack:main"
"lelab-frontend" = "scripts.frontend:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scripts*"]

[tool.setuptools.package-data]
app = ["static/**/*"]
